conform to the node 6 Buffer.indexOf api
Tested against adapted Buffer.indexOf test suite from node 6. This suite is pretty extensive but had a few rough areas around undefined behaviors, particularly with providing a buffer and a string encoding. I've left that particular test commented out.

The one weak area of the tests are with proving an encoding and not an offset. Only a handful of tests cover this and it's not declared that's what's being tested.

Fixes 
 - argument errors
 - incorrectly matching 0 length searches
 - incorrectly handling negative and infinite offsets
 - incorrect handling of non buffer searches

Adds 
 - search encoding for strings and byte declarations